Refactors store into Recipe Store and actions into RecipeActions
